Wave 25: 66 functions, 1696 -> 1762, and sub_080152C0 unblocked after thirteen waves
Three agents, one address block each (0x08034, 0x08044, 0x08015), 42
assigned and all 42 matched; 66 promoted after extension work. Both builds
reproduce 14dd0b22c894865867aff89e8116b2dffae25605. Index metric
18.72% -> 19.26% (109,328 of 567,538 bytes).

The address-locality screen added last commit produced this wave's
batching directly, on its first real use. It is the batching list now.

sub_080152C0 IS MATCHED, and it had been blocked since wave 12. The
recorded reason was that its `s32` first parameter was wrong and fixing it
needed an edit to src/proc.c, which this project never touches. That was
wrong on both halves. The parameter is not a problem to solve: all eight
call sites already pass a pointer through an explicit `(s32)` cast, which
is byte-neutral, and the body casts it back. The blocker was the RETURN
type -- `void` where the ROM returns `s8` -- and it can be fixed alone,
because every caller discards the result and a discarded return emits no
narrowing. src/proc.c is not edited by the change; one header line moves.

W25-C found this, proved it byte-neutral, and then DID NOT APPLY IT --
it escalated instead, because proc.c is a multi-function file that only a
full `make SPLIT=1 compare` can gate. That was the right call and it is
the behaviour the brief asks for.

## A wrong turn worth recording, because it cost six working functions

Three agents declared narrow return types (`u8` sub_0804415C, `bool8`
sub_0803B628, `u32` sub_0808B6B0) against promoted definitions saying
`int`. tools/proto_check.py flagged all three -- see below -- and the
reconciliation went the wrong way: narrowing each DEFINITION was probed,
all three appeared to break, so the prototypes were "corrected" to `int`.
That broke six already-verified functions, every one a caller.

The probe was invalid. `trymatch` compiles `work/<fn>/<fn>.c`, NOT
src/decomp/c_<addr>.c, so editing the promoted file and re-running it
measures the STALE DRAFT. With `python tools/sync_work.py <fn>` in
between, all three narrow types match byte-for-byte. The agents were
right; their types are restored on both sides.

The rule underneath is wave 24's and this is its second confirmation: a
return type is settled by the CALLERS, not the callee. The callee's body
compiles byte-identically either way -- that is what lets a wrong type
survive -- while sub_0804423C's `if (sub_0804415C(a1))` came out exactly
4 bytes SHORT with an `int` return, which is the narrowing the ROM has.
Corollary: a header-vs-definition conflict does not mean the header is
wrong. "The definition wins" makes the build link; it is a tie-break, not
evidence. Written up in docs/agbcc-codegen.md.

## proto_check.py was reading half of every signature

It compared parameter lists and never return types, so it reported clean
on the three conflicts above -- the permanently-green twin of the
permanently-red self-test. It now compares both and says which half
differs. Four prototype conflicts landed in this one wave
(sub_08044080's parameters plus those three returns), every one invisible
to per-function trymatch and every one a guaranteed split-build failure.
Corpus-wide it is clean at 739 prototypes.

## Other findings

sub_08034A44's block: gUnknown_03004780 retyped u16 -> s16 on an `ldrsh`
read, which reaches 16 promoted units in 0x0805C-0x08060. Regression-swept
all 17 functions in them: clean. A movs-sized constant stored with `strh`
is byte-identical either way.

sub_0803442C was 8 bytes SHORT and closed as one BINDING LOCAL too few --
wave 23's diagnostic, applied directly and correctly.

sub_080344F0 is parked at 68 bytes, size-exact.

Two sweep hazards, both now handled in the sweep itself: `start_function`
SCAFFOLDS (18 corpus-wide) compile to 4 bytes and read as catastrophic
failures although nobody ever worked them, and parked entries must be
excluded or they are re-reported every wave. And CRLF again -- Python
text-mode writes the promotion list with \r\n, `tr '\n' ' '` leaves the \r
glued to each name, and promote.py then reports a function "not in the
index" that is plainly in it.

+++ src/decomp/c_0804415C.c
@@ -14,7 +14,7 @@
  * sub_0804415C @ 0x0804415C
  */
 
-int sub_0804415C(int a1)
+u8 sub_0804415C(int a1)
 {
     return gUnknown_08499598[a1].unk1e != 0;
 }
